pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tmrc"
version = "0.1.0"
description = "Transformer Model Research Codebase (TMRC)"
authors = [
{ name = "Research and Engineering at Kempner Institute", email = "kempner-research-engineering@g.harvard.edu" }
]
license = { file = "LICENSE" }
classifiers = [
"Programming Language :: Python :: 3",
"License :: OSI Approved :: MIT License",
"Operating System :: OS Independent"
]
dependencies = [
    "filelock==3.16.1",
    "tatm @ git+https://github.com/KempnerInstitute/tatm.git@main",
    "fsspec==2024.6.1",
    "jinja2==3.1.4",
    "markupsafe==2.1.5",
    "mpmath==1.3.0",
    "networkx==3.2.1",
    "numpy==2.0.2",
    "sympy==1.13.1",
    "torch==2.5.0",
    "triton",
    "typing-extensions==4.12.2",
    "hydra-core==1.3.2"

]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-sphinx",
    "twine>=1.11.0",
    "setuptools",
    "wheel",
    "build",
]
docs = [
    "sphinx",
    "sphinx-rtd-theme",
    "sphinxcontrib-napoleon",
    "myst-parser",
    "nbsphinx",
    "recommonmark",
    "sphinx-copybutton",
    "sphinxcontrib-bibtex",

]